Allow dragging tiles onto the layout view to change the track arrangement. * Add method for editing EditableScenario. * Add new DragDropView as a view containing the tiles that can be dragged into the layout view. * Break out the track-drawing code in CocoaLayoutView into a separate TrackDrawer class.  The new class allows the track to be drawn in other contexts (such as when creating the icons in the DragDropView). * Add machinery for allowing drags from the DragDropView to the CocoaLayoutView.

+++ SwitchTower/Scenario.h
@@ -78,8 +78,8 @@
 @property (nonatomic, retain) NSArray *all_trains;
 @property (nonatomic, retain) NSMutableArray *cellLengths;
 @property (nonatomic, retain) NSArray *timetableNames;
-@property (nonatomic) NSUInteger tileColumns;
-@property (nonatomic) NSUInteger tileRows;
+@property (nonatomic) int tileColumns;
+@property (nonatomic) int tileRows;
 // TODO(bowdidge): Make more structured and validated.
 @property (nonatomic, retain) NSArray *tileStrings;
 @property (nonatomic, retain) NSString *helpString;
